[DF-7335] Warnings fix for windows x64
Former-commit-id: f0b18c41eb1c4ef665ce81a6da982edaa766f517

+++ Sources/Internal/Debug/Profiler.cpp
@@ -383,13 +383,13 @@ fltDec( float f )
 static void
 DumpInternal( const std::vector<CounterInfo>& result, bool showPercents=false )
 {
-    unsigned    max_name_len = 0;
+    size_t max_name_len = 0;
     
     for( size_t i=0,i_end=result.size(); i!=i_end; ++i )
     {
         uint32  pi      = result[i].parentIndex;
         uint32  indent  = 0;
-        uint32  len     = 0;
+        size_t  len     = 0;
 
         while( pi != InvalidIndex )
         {
@@ -419,7 +419,7 @@ DumpInternal( const std::vector<CounterInfo>& result, bool showPercents=false )
             ++indent;
         }
 
-        int  text_len = 0;       
+        size_t text_len = 0;       
 
         if( result[i].name )
             text_len = Snprintf( text+indent*2, sizeof(text)-indent*2, "%s", result[i].name );
